* fixed login error (required to press the login button 2x) when the server was already running. * made the error messages disappear in release mode.

+++ src/PythonObjects/PyObjects.cpp
@@ -654,7 +654,7 @@ std::string PyTuple::GetString( const int index )
 
 bool PyTuple::GetString( const int index, std::string & str )
 {
-	if (index > mTuple.size())
+	if (index > (int)mTuple.size())
 		return false;
 
 	PyChameleon * itr = mTuple[index];
